test: renamed tests and data files
