--- __tests__/data/api/error.data.js
@@ -1,5 +1,0 @@
-export const authError = {
-  message: 'Requires authentication',
-  documentation_url:
-    'https://developer.github.com/v3/activity/notifications/#list-your-notifications',
-};
